Refactor: Break up BusRouteDetail into components. Rename RouteInfo page to RouteStats.
